# go-expert/09-api/docs/swagger.yaml
basePath: /v1
definitions:
  database.Pageable:
    properties:
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalElements:
        type: integer
      totalPages:
        type: integer
    type: object
  error.Problem:
    properties:
      detail:
        type: string
      status:
        type: integer
      timestamp:
        type: string
      title:
        type: string
    type: object
  product.PageProductOutputDTO:
    properties:
      content:
        items:
          $ref: '#/definitions/product.ProductOutPutDTO'
        type: array
      pageable:
        $ref: '#/definitions/database.Pageable'
    type: object
  product.ProductDtoInputDTO:
    properties:
      name:
        type: string
      price:
        type: number
    type: object
  product.ProductOutPutDTO:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  user.UserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  user.UserLoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  user.UserToken:
    properties:
      access_token:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: atendimento@fullcycle.com.br
    name: Wesley Willians
    url: http://www.fullcycle.com.br
  description: Product API with auhtentication
  license:
    name: Full Cycle License
    url: http://www.fullcycle.com.br
  termsOfService: http://swagger.io/terms/
  title: Go Expert API Example
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: get all products
      parameters:
      - description: page number
        in: query
        name: page
        type: string
      - description: page size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/product.PageProductOutputDTO'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.Problem'
      security:
      - ApiKeyAuth: []
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create products
      parameters:
      - description: product request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.ProductDtoInputDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.Problem'
      security:
      - ApiKeyAuth: []
      summary: Create product
      tags:
      - products
  /users:
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.Problem'
      summary: Create user
      tags:
      - users
  /users/generate_token:
    post:
      consumes:
      - application/json
      description: Get a user JWT
      parameters:
      - description: user credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserToken'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.Problem'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.Problem'
      summary: Get a user JWT
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"